Conflicts:
	CHANGES
	cpp/demo/Freeze/backup/.depend
	cpp/demo/Freeze/bench/.depend
	cpp/demo/Freeze/casino/.depend
	cpp/demo/Freeze/customEvictor/.depend
	cpp/demo/Freeze/library/.depend
	cpp/demo/Freeze/phonebook/.depend
	cpp/demo/Freeze/transform/.depend
	cpp/demo/Glacier2/callback/.depend
	cpp/demo/Glacier2/chat/.depend
	cpp/demo/Ice/async/.depend
	cpp/demo/Ice/bidir/.depend
	cpp/demo/Ice/callback/.depend
	cpp/demo/Ice/converter/.depend
	cpp/demo/Ice/hello/.depend
	cpp/demo/Ice/invoke/.depend
	cpp/demo/Ice/latency/.depend
	cpp/demo/Ice/minimal/.depend
	cpp/demo/Ice/multicast/.depend
	cpp/demo/Ice/nested/.depend
	cpp/demo/Ice/nrvo/.depend
	cpp/demo/Ice/session/.depend
	cpp/demo/Ice/throughput/.depend
	cpp/demo/Ice/value/.depend
	cpp/demo/IceBox/hello/.depend
	cpp/demo/IceGrid/allocate/.depend
	cpp/demo/IceGrid/icebox/.depend
	cpp/demo/IceGrid/replication/.depend
	cpp/demo/IceGrid/sessionActivation/.depend
	cpp/demo/IceGrid/simple/.depend
	cpp/demo/IceStorm/clock/.depend
	cpp/demo/IceStorm/counter/.depend
	cpp/demo/IceStorm/replicated/.depend
	cpp/demo/IceStorm/replicated2/.depend
	cpp/demo/book/freeze_filesystem/.depend
	cpp/demo/book/lifecycle/.depend
	cpp/demo/book/printer/.depend
	cpp/demo/book/simple_filesystem/.depend
	cpp/src/Freeze/.depend
	cpp/src/FreezeScript/.depend
	cpp/src/Ice/.depend
	cpp/src/Ice/UdpTransceiver.cpp
	cpp/src/Ice/UdpTransceiver.h
	cpp/src/IceBox/.depend
	cpp/src/IceGrid/.depend
	cpp/src/IceGridLib/.depend
	cpp/src/IcePatch2/.depend
	cpp/src/IceStorm/.depend
	cpp/src/slice2freeze/.depend
	cpp/test/Freeze/complex/.depend
	cpp/test/Freeze/dbmap/.depend
	cpp/test/Freeze/evictor/.depend
	cpp/test/Freeze/oldevictor/.depend
	cpp/test/FreezeScript/dbmap/.depend
	cpp/test/FreezeScript/evictor/.depend
	cpp/test/Glacier2/attack/.depend
	cpp/test/Glacier2/dynamicFiltering/.depend
	cpp/test/Glacier2/router/.depend
	cpp/test/Glacier2/sessionControl/.depend
	cpp/test/Glacier2/ssl/.depend
	cpp/test/Glacier2/staticFiltering/.depend
	cpp/test/Ice/adapterDeactivation/.depend
	cpp/test/Ice/background/.depend
	cpp/test/Ice/binding/.depend
	cpp/test/Ice/checksum/.depend
	cpp/test/Ice/checksum/server/.depend
	cpp/test/Ice/custom/.depend
	cpp/test/Ice/exceptions/.depend
	cpp/test/Ice/facets/.depend
	cpp/test/Ice/faultTolerance/.depend
	cpp/test/Ice/gc/.depend
	cpp/test/Ice/hold/.depend
	cpp/test/Ice/inheritance/.depend
	cpp/test/Ice/interceptor/.depend
	cpp/test/Ice/location/.depend
	cpp/test/Ice/objects/.depend
	cpp/test/Ice/operations/.depend
	cpp/test/Ice/proxy/.depend
	cpp/test/Ice/retry/.depend
	cpp/test/Ice/servantLocator/.depend
	cpp/test/Ice/slicing/exceptions/.depend
	cpp/test/Ice/slicing/objects/.depend
	cpp/test/Ice/stream/.depend
	cpp/test/Ice/stringConverter/.depend
	cpp/test/Ice/timeout/.depend
	cpp/test/Ice/udp/.depend
	cpp/test/IceBox/configuration/.depend
	cpp/test/IceGrid/activation/.depend
	cpp/test/IceGrid/allocation/.depend
	cpp/test/IceGrid/deployer/.depend
	cpp/test/IceGrid/distribution/.depend
	cpp/test/IceGrid/replicaGroup/.depend
	cpp/test/IceGrid/replication/.depend
	cpp/test/IceGrid/session/.depend
	cpp/test/IceGrid/simple/.depend
	cpp/test/IceGrid/update/.depend
	cpp/test/IceSSL/configuration/.depend
	cpp/test/IceStorm/federation/.depend
	cpp/test/IceStorm/federation2/.depend
	cpp/test/IceStorm/rep1/.depend
	cpp/test/IceStorm/repgrid/.depend
	cpp/test/IceStorm/repstress/.depend
	cpp/test/IceStorm/single/.depend
	cpp/test/IceStorm/stress/.depend
	cpp/test/Slice/keyword/.depend
	cs/src/Ice/Instance.cs
	cs/src/IceSSL/ConnectorI.cs
	java/demo/book/simple_filesystem/Filesystem/DirectoryI.java
	java/demo/book/simple_filesystem/Filesystem/FileI.java
	java/src/IceInternal/TcpConnector.java
	java/src/IceSSL/ConnectorI.java
	py/modules/IcePy/.depend
	rb/src/IceRuby/.depend

+++ cpp/src/IceGrid/RegistryI.cpp
@@ -167,6 +167,25 @@ RegistryI::~RegistryI()
 bool
 RegistryI::start()
 {
+    try
+    {
+        if(!startImpl())
+        {
+            stop();
+            return false;
+        }
+    }
+    catch(...)
+    {
+        stop();
+        throw;
+    }
+    return true;
+}
+
+bool
+RegistryI::startImpl()
+{
     assert(_communicator);
     PropertiesPtr properties = _communicator->getProperties();
 
@@ -687,7 +706,10 @@ RegistryI::stop()
     // ensure that there will be no more invocations on IceStorm once
     // it's shutdown.
     //
-    _database->destroyTopics();
+    if(_database)
+    {
+        _database->destroyTopics();
+    }
 
     try
     {
@@ -1028,6 +1050,9 @@ RegistryI::getPermissionsVerifier(const ObjectAdapterPtr& adapter,
         {
             try
             {
+#if defined(__BCPLUSPLUS__) && (__BCPLUSPLUS__ >= 0x0600)
+                IceUtil::DummyBCC dummy;
+#endif
                 verifier = _communicator->propertyToProxy(verifierProperty);
             }
             catch(const ProxyParseException&)
@@ -1110,6 +1135,9 @@ RegistryI::getPermissionsVerifier(const ObjectAdapterPtr& adapter,
     Glacier2::PermissionsVerifierPrx verifierPrx;
     try
     {
+#if defined(__BCPLUSPLUS__) && (__BCPLUSPLUS__ >= 0x0600)
+        IceUtil::DummyBCC dummy;
+#endif
         //
         // Set the permission verifier proxy locator to the internal
         // locator. We can't use the "public" locator, this could lead
@@ -1151,6 +1179,9 @@ RegistryI::getSSLPermissionsVerifier(const IceGrid::LocatorPrx& locator, const s
         {
             try
             {
+#if defined(__BCPLUSPLUS__) && (__BCPLUSPLUS__ >= 0x0600)
+                IceUtil::DummyBCC dummy;
+#endif
                 verifier = _communicator->propertyToProxy(verifierProperty);
             }
             catch(const ProxyParseException&)
@@ -1194,6 +1225,9 @@ RegistryI::getSSLPermissionsVerifier(const IceGrid::LocatorPrx& locator, const s
     Glacier2::SSLPermissionsVerifierPrx verifierPrx;
     try
     {
+#if defined(__BCPLUSPLUS__) && (__BCPLUSPLUS__ >= 0x0600)
+        IceUtil::DummyBCC dummy;
+#endif
         //
         // Set the permission verifier proxy locator to the internal
         // locator. We can't use the "public" locator, this could lead
